Consistify namespacing conventions
