+++ app/workers/search_for_external_registry_bikes_worker.rb
@@ -1,6 +1,6 @@
 class SearchForExternalRegistryBikesWorker < ApplicationWorker
   def perform(serial_number)
-    client = ExternalRegistry::VerlorenOfGevondenClient.new
+    client = ExternalRegistryClient::VerlorenOfGevondenClient.new
     client.search(serial_number)
   end
 end
