+++ spec/factories/external_registry_bikes.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :external_registry_bike,
-          class: "ExternalRegistryBikes::VerlorenOfGevondenBike" do
+          class: "ExternalRegistryBike::VerlorenOfGevondenBike" do
     external_id { 10.times.map { rand(10) }.join }
     serial_number { 10.times.map { rand(10) }.join }
     country { Country.netherlands }
